Implement Model and Controller + Basic Text Formatting Features
In this commit, the model and controller parts were implemented. The controller have references to the view and model, and it attaches Swing Actions to the different elements of the View. In terms of architecture, the objective is to keep as close as possible to an MVC architecture. One important note is that the Controller will mainly work as the Brain of the program, and View will display the Model. Model and View are not aware of each other. Although this should have been in a separate commit, the basic features of text editing, such as Bold, Italic, Underline, Subscript, Superscript were implemented.