Creation of the JDialog for Search Tool and Addition of ScaledEditorKit
The custom non-modal dialog was added and temporarily binded to the corresponding menu item in order to test the visual. Additionally, the ScaledEdtitorKit (from java-sl.com) was added to experiment and integrate into the function. It is important to note that the code was made in Java 8, and some functions do not behave correctly in the newer version specially due to how the reflection is handled differently. With this current configuration, the Justify function gets disabled as effect.