Addition of Open/Save Document
This is implemented by serializing the DefaultStyledDocument into RTF and saving it into an RTF file. And the deserialization works the same on the opposite way. It is worth noting that this is not ideal, as the RTF used by Swing lacks support for newer features, but it is enough as a start point.